+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageListHelperViews.swift
@@ -199,4 +199,9 @@ extension View {
             return Color(colors.messageOtherUserTextColor)
         }
     }
+    
+    func textColor(currentUser: Bool) -> Color {
+        @Injected(\.colors) var colors
+        return currentUser ? Color(colors.messageCurrentUserTextColor) : Color(colors.messageOtherUserTextColor)
+    }
 }