+++ Sources/StreamChatSwiftUI/ColorPalette.swift
@@ -79,6 +79,11 @@ public struct ColorPalette {
     public lazy var reactionCurrentUserColor: UIColor? = UIColor(tintColor)
     public lazy var reactionOtherUserColor: UIColor? = textLowEmphasis
     public lazy var selectedReactionBackgroundColor: UIColor? = nil
+    public var voiceMessageCurrentUserBackground: UIColor = .streamInnerBorder
+    public var voiceMessageOtherUserBackground: UIColor = .streamBarsBackground
+    public var voiceMessageCurrentUserRecordingBackground: UIColor = .streamBarsBackground
+    public var voiceMessageOtherUserRecordingBackground: UIColor = .streamBarsBackground
+    public var voiceMessageControlBackground: UIColor = .streamWhiteStatic
 
     // MARK: - Composer
 
@@ -114,6 +119,7 @@ private extension UIColor {
     static let streamInnerBorder = mode(0xdbdde1, 0x272a30)
     static let streamHighlight = mode(0xfbf4dd, 0x333024)
     static let streamDisabled = mode(0xb4b7bb, 0x4c525c)
+    static let streamBarsBackground = mode(0xffffff, 0x17191c)
 
     // Currently we are not using the correct shadow color from figma's color palette. This is to avoid
     // an issue with snapshots inconsistency between Intel vs M1. We can't use shadows with transparency.